Dbg statements and proc_kill

+++ kernel/proc/proc.c
@@ -133,7 +133,12 @@ proc_create(char *name)
 	new_proc->p_comm[PROC_NAME_LEN - 1] = '\0';
 	list_init(&(new_proc->p_threads)); 	/* initialize list  to track list of threads */
 	list_init(&(new_proc->p_children)); /* initialize list to track list of children */
-	new_proc->p_pproc = (PID_IDLE == pid) ? NULL : curproc; /* parent process is the current process that's running */
+
+	if(PID_IDLE == pid) { dbg(DBG_PRINT, "(GRADING1A)\n");
+		new_proc->p_pproc = NULL; /* parent process is the current process that's running */
+	}else { dbg(DBG_PRINT, "(GRADING1A)\n");
+		new_proc->p_pproc = curproc;
+	}
 	new_proc->p_status = NULL;
 	new_proc->p_state = PROC_RUNNING;
 	sched_queue_init(&(new_proc->p_wait)); 		/* initialize wait queue */
@@ -143,12 +148,14 @@ proc_create(char *name)
 	list_link_init(&(new_proc->p_list_link));
 	list_link_init(&(new_proc->p_child_link));
 	list_insert_tail(&_proc_list, &(new_proc->p_list_link));
-	if (NULL != curproc)
+	if (NULL != curproc) { dbg(DBG_PRINT, "(GRADING1A)\n");
 		list_insert_tail(&(curproc->p_children), &(new_proc->p_child_link)); /* for idle process, there is no curproc */
+	}
 
 	/* set initproc global variable */
-	if (PID_INIT == new_proc->p_pid)
+	if (PID_INIT == new_proc->p_pid) { dbg(DBG_PRINT, "(GRADING1A)\n");
 		proc_initproc = new_proc;
+	}
 
 	/* VFS-related: */
 	int index = 0;
@@ -292,38 +299,16 @@ proc_kill(proc_t *p, int status)
 	dbg(DBG_PRINT, "INFO : place cancel request on the process's threads in case required\n");
 	/* iterate over threads and cancel them */
 	kthread_t *thr;
-	int is_any_thread_alive = 0;
 	list_iterate_begin(&p->p_threads, thr, kthread_t, kt_plink)
 	{
 		KASSERT(NULL != thr);
-		/*if(KT_SLEEP == thr->kt_state || KT_SLEEP_CANCELLABLE == thr->kt_state) {
-			is_any_thread_alive = 1;
-			thr->kt_cancelled = 1;
-			sched_wakeup_on(thr->kt_wchan);
-		}*/
 		if(KT_EXITED != thr->kt_state || KT_NO_STATE != thr->kt_state){ /* thread is in waitQ or runQ */
 			dbg(DBG_PRINT, "INFO : placing the cancel request on process's thread\n");
-			is_any_thread_alive = 1;
 			kthread_cancel(thr, (void*)-1); /* cancel the thread */
 		}
 		/* this thread need not be removed from the thread list now. it can be handled in the do_waitpid */
 	}list_iterate_end();
 
-	if(!is_any_thread_alive) {
-		dbg(DBG_PRINT, "INFO : killing  process = %d, parent PID = %d\n", p->p_pid, p->p_pproc->p_pid);
-		p->p_status = status; 		/* set the status for the current process, this will be returned to the parent when it calls do_waitpid() */
-		p->p_state = PROC_DEAD; 		/* mark the process is DEAD */
-		KASSERT(NULL != &(p->p_pproc->p_wait));
-		dbg(DBG_PRINT, "INFO : waking up the parent %d of process %d\n", p->p_pproc->p_pid, p->p_pid);
-		sched_wakeup_on(&(p->p_pproc->p_wait)); /* wake up the parent process it may wait for the child to die */
-	}
-	/*
-	p->p_status = status; 	 //set the status for the process
-	p->p_state = PROC_DEAD;  //set the state for the process
-
-	KASSERT(NULL != &p->p_pproc->p_wait);
-	sched_wakeup_on(&p->p_pproc->p_wait);
-	*/
 }
 
 /*
@@ -342,7 +327,7 @@ proc_kill_all()
 	{
 		KASSERT(NULL != child);
 		/* kill all the process except IDLE and direct children of IDLE */
-		if (curproc != child && PID_IDLE != child->p_pid && PID_IDLE != child->p_pproc->p_pid) {
+		if (PROC_DEAD != child->p_state && curproc != child && PID_IDLE != child->p_pid && PID_IDLE != child->p_pproc->p_pid) {
 			dbg(DBG_PRINT, "INFO : kill process %d\n", child->p_pid);
 			proc_kill(child, child->p_status);
 		}
@@ -394,10 +379,10 @@ void proc_cleanup_memory(proc_t* dead_proc) {
 
 	/* clean up process space */
 	dbg(DBG_PRINT, "INFO : removing process links\n");
-	if (list_link_is_linked(&dead_proc->p_list_link))
-		list_remove(&dead_proc->p_list_link); 	/* remove child from the global list */
-	if (list_link_is_linked(&dead_proc->p_child_link))
-		list_remove(&dead_proc->p_child_link); /* remove child from parents(curproc) child list */
+	if (list_link_is_linked(&dead_proc->p_list_link)){ dbg(DBG_PRINT, "(GRADING1A)\n");
+		list_remove(&dead_proc->p_list_link); 	/* remove child from the global list */ }
+	if (list_link_is_linked(&dead_proc->p_child_link)) { dbg(DBG_PRINT, "(GRADING1A)\n");
+		list_remove(&dead_proc->p_child_link); /* remove child from parents(curproc) child list */ }
 	KASSERT(NULL != dead_proc->p_pagedir);  /* this process should have a valid pagedir */
 	dbg(DBG_PRINT, "(GRADING1A 2.c)\n");
 	dbg(DBG_PRINT, "INFO : removing page directory\n");
@@ -433,7 +418,7 @@ do_waitpid(pid_t pid, int options, int *status)
 	KASSERT(NULL != curproc && NULL!= &(curproc->p_children));
 	KASSERT((NULL != pid && pid>0) || (-1 == pid));
 	KASSERT(0 == options);
-	if (list_empty(&curproc->p_children)) {
+	if (list_empty(&curproc->p_children)) { dbg(DBG_PRINT, "(GRADING1A)\n");
 		dbg(DBG_PRINT, "INFO : no children found, so returns -ECHILD \n");
 		if(status) *status = -1;
 		return -ECHILD;
@@ -447,9 +432,9 @@ do_waitpid(pid_t pid, int options, int *status)
 	wait_pid: /* label for goto */
 	list_iterate_begin(&curproc->p_children, child, proc_t, p_child_link)
 	{
-		if (-1 == pid) { /* user is not interested in specific pid */
+		if (-1 == pid) { dbg(DBG_PRINT, "(GRADING1A)\n"); /* user is not interested in specific pid */
 			pid_found = 1;
-			if (PROC_DEAD == child->p_state) { /* any of the child process is dead */
+			if (PROC_DEAD == child->p_state) { dbg(DBG_PRINT, "(GRADING1A)\n"); /* any of the child process is dead */
 				found_dead_child = 1;
 				dead_child = child;
 				break; /* once we found the child that's dead */
@@ -474,22 +459,26 @@ do_waitpid(pid_t pid, int options, int *status)
 
 	if (0 == pid_found) {
 		dbg(DBG_PRINT, "INFO : given pid(%d) is not found in the curproc's child list\n", pid);
-		if(status) *status = -1;
+		if(status) {
+			*status = -1;
+		}
 		return -ECHILD; /*given PID couldnt be found from the curpocess child list */
 	}
-	if (0 == found_dead_child) { /* Process is fond and it is not dead, wait for it till it dies */
+	if (0 == found_dead_child) { dbg(DBG_PRINT, "(GRADING1A)\n"); /* Process is fond and it is not dead, wait for it till it dies */
 		dbg(DBG_PRINT, "INFO : none of the child processes of the given process(PID = %d) is not dead yet. so, it goes for sleep\n", curproc->p_pid);
 		sched_sleep_on(&curproc->p_wait);
 		dbg(DBG_PRINT, "INFO : process %d woke up \n", curproc->p_pid);
 		goto wait_pid;
-	} else { /* found_dead_child ==  1*/
+	} else { dbg(DBG_PRINT, "(GRADING1A)\n"); /* found_dead_child ==  1*/
 		KASSERT(NULL != dead_child);  /* the dead child process should not be NULL */
 		dbg(DBG_PRINT, "(GRADING1A 2.c)\n");
 		pid_t dead_child_pid = dead_child->p_pid;
 		KASSERT(-1 == dead_child_pid || dead_child->p_pid == dead_child_pid); /* should be able to find a valid process ID for the process */
 		dbg(DBG_PRINT, "INFO : found dead child %d\n", dead_child_pid);
 		dbg(DBG_PRINT, "(GRADING1A 2.c)\n");
-		if(status) *status = dead_child->p_status;
+		if(status) { dbg(DBG_PRINT, "(GRADING1A)\n");
+			*status = dead_child->p_status;
+		}
 
 		proc_cleanup_memory(dead_child);
 		return dead_child_pid;
